Split off ALC tests
